Use configured endpoints consistently
- remove duplication in ResetController
- use adminpaneluser base path + slugs to create paths
- move unique field validation to Sugar
- remove `public` from handleValidationError functions
- split up AdminPanelUser protocol conformance to more separate files
