+++ Sources/AdminPanel/Configs/TagTemplatePaths.swift
@@ -1,5 +1,6 @@
 /// Configuration for template paths used when rendering tags.
 public struct TagTemplatePaths {
+
     /// Path to template for Quill WYSIWYG field.
     public let wysiwygField: String
