+++ Sources/AdminPanel/Models/AdminPanelUserRole/AdminPanelUserRoleType.swift
@@ -1,0 +1,3 @@
+public protocol AdminPanelUserRoleType: LosslessStringConvertible, Comparable, Codable {
+    var menuPath: String { get }
+}
