+++ Sources/AdminPanel/Providers/AdminPanelProvider.swift
@@ -34,22 +34,24 @@ public final class AdminPanelProvider<U: AdminPanelUserType>: Provider {
             CurrentURLMiddleware()
         ]
 
+        let endpoints = config.endpoints
         let secure: [Middleware] = unsecure + [
-            RedirectMiddleware<U>(path: AdminPanelEndpoints.default.login),
-            ShouldResetPasswordMiddleware<U>(path: AdminPanelEndpoints.default.renderEditMe),
+            RedirectMiddleware<U>(path: config.endpoints.login),
+            ShouldResetPasswordMiddleware<U>(path:
+                "\(endpoints.adminPanelUserBasePath)/\(endpoints.meSlug)/\(endpoints.editSlug)"
+            ),
             CurrentUserMiddleware<U>()
         ]
 
         self.middlewares = .init(
-            unsecure: unsecure,
-            secure: secure
+            secure: secure,
+            unsecure: unsecure
         )
     }
 
     /// See Service.Provider.Register
     public func register(_ services: inout Services) throws {
         try services.register(AuthenticationProvider())
-        try services.register(BootstrapProvider())
         try services.register(CurrentURLProvider())
         try services.register(CurrentUserProvider<U>())
         try services.register(FlashProvider())
@@ -59,13 +61,36 @@ public final class AdminPanelProvider<U: AdminPanelUserType>: Provider {
             config: .init(
                 name: config.name,
                 baseURL: config.baseURL,
-                endpoints: ResetEndpoints(
-                    renderResetPasswordRequest: "/admin/users/reset-password/request",
-                    resetPasswordRequest: "/admin/users/reset-password/request",
-                    renderResetPassword: "/admin/users/reset-password",
-                    resetPassword: "/admin/users/reset-password"
-                ),
+                endpoints: config.resetEndpoints,
                 signer: config.resetSigner,
+                responses: ResetResponses(
+                    resetPasswordRequestForm: { [config] req in
+                        try req
+                            .view()
+                            .render(config.views.login.requestResetPassword, on: req)
+                            .encode(for: req)
+                    },
+                    resetPasswordUserNotified: { [config] req in
+                        req.future(req
+                            .redirect(to: config.endpoints.login)
+                            .flash(.success, "Email with reset link sent.")
+                        )
+                    },
+                    resetPasswordForm: { [config] req, user in
+                        try req.addFields(forType: U.self)
+                        return try req
+                            .view()
+                            .render(config.views.login.resetPassword, on: req)
+                            .encode(for: req)
+                    },
+                    resetPasswordSuccess: { [config] req, user in
+                        req.future(req
+                            // TODO: make configurable
+                            .redirect(to: config.endpoints.login)
+                            .flash(.success, "Your password has been updated.")
+                        )
+                    }
+                ),
                 controller: AdminPanel.ResetController<U>()
             )
         ))
@@ -79,7 +104,7 @@ public final class AdminPanelProvider<U: AdminPanelUserType>: Provider {
         ))
         services.register(config)
         services.register(KeyedCacheSessions.self)
-        services.register(self.middlewares)
+        services.register(middlewares)
     }
 
     /// See Service.Provider.boot
