+++ Sources/AdminPanel/routes.swift
@@ -9,27 +9,33 @@ public struct AdminPanelEndpoints {
     public let login: String
     public let logout: String
     public let dashboard: String
-    public let renderAdminPanelUserList: String
-    public let renderCreateAdminPanelUser: String
-    public let createAdminPanelUser: String
-    public let renderEditMe: String
+    public let adminPanelUserBasePath: String
+
+    public let createSlug: String
+    public let deleteSlug: String
+    public let editSlug: String
+    public let meSlug: String
 
     public init(
+        adminPanelUserBasePath: String = "/admin/users",
+        dashboard: String = "/admin",
         login: String = "/admin/login",
         logout: String = "/admin/logout",
-        dashboard: String = "/admin",
-        renderAdminPanelUserList: String = "/admin/users",
-        renderCreateAdminPanelUser: String = "/admin/users/create",
-        createAdminPanelUser: String = "/admin/users/create",
-        renderEditMe: String = "/admin/users/me/edit"
+
+        createSlug: String = "create",
+        deleteSlug: String = "delete",
+        editSlug: String = "edit",
+        meSlug: String = "me"
     ) {
         self.login = login
         self.logout = logout
         self.dashboard = dashboard
-        self.renderAdminPanelUserList = renderAdminPanelUserList
-        self.renderCreateAdminPanelUser = renderCreateAdminPanelUser
-        self.createAdminPanelUser = createAdminPanelUser
-        self.renderEditMe = renderEditMe
+        self.adminPanelUserBasePath = adminPanelUserBasePath
+
+        self.createSlug = createSlug
+        self.deleteSlug = deleteSlug
+        self.editSlug = editSlug
+        self.meSlug = meSlug
     }
 
     public static var `default`: AdminPanelEndpoints {
@@ -38,8 +44,8 @@ public struct AdminPanelEndpoints {
 }
 
 public struct AdminPanelMiddlewares: Service {
-    public let unsecure: [Middleware]
     public let secure: [Middleware]
+    public let unsecure: [Middleware]
 }
 
 public extension Router {
@@ -50,61 +56,70 @@ public extension Router {
         let config: AdminPanelConfig<U> = try container.make()
         let middlewares: AdminPanelMiddlewares = try container.make()
 
-        let unprotected = self.grouped(middlewares.unsecure)
-        let protected = self.grouped(middlewares.secure)
+        let unprotected = grouped(middlewares.unsecure)
+        let protected = grouped(middlewares.secure)
+
+        let endpoints = config.endpoints
+        let controllers = config.controllers
 
         // MARK: Login routes
 
-        unprotected.get(config.endpoints.login, use: config.controllers.loginController.renderLogin)
-        unprotected.post(config.endpoints.login, use: config.controllers.loginController.login)
-        unprotected.get(config.endpoints.logout, use: config.controllers.loginController.logout)
+        unprotected.get(endpoints.login, use: controllers.loginController.renderLogin)
+        unprotected.post(endpoints.login, use: controllers.loginController.login)
+        unprotected.get(endpoints.logout, use: controllers.loginController.logout)
 
         // MARK: Dashboard routes
 
         protected.get(
-            config.endpoints.dashboard,
-            use: config.controllers.dashboardController.renderDashboard
+            endpoints.dashboard,
+            use: controllers.dashboardController.renderDashboard
         )
 
         // MARK: Admin Panel User routes
 
         protected.get(
-            config.endpoints.renderAdminPanelUserList,
-            use: config.controllers.adminPanelUserController.renderList
+            endpoints.adminPanelUserBasePath,
+            use: controllers.adminPanelUserController.renderList
         )
         protected.get(
-            config.endpoints.renderCreateAdminPanelUser,
-            use: config.controllers.adminPanelUserController.renderCreate
+            endpoints.adminPanelUserBasePath,
+            endpoints.createSlug,
+            use: controllers.adminPanelUserController.renderCreate
         )
         protected.post(
-            config.endpoints.createAdminPanelUser,
-            use: config.controllers.adminPanelUserController.create
+            endpoints.adminPanelUserBasePath,
+            endpoints.createSlug,
+            use: controllers.adminPanelUserController.create
         )
         protected.get(
-            "/admin/users",
+            endpoints.adminPanelUserBasePath,
             U.parameter,
-            "edit",
-            use: config.controllers.adminPanelUserController.renderEditUser
+            endpoints.editSlug,
+            use: controllers.adminPanelUserController.renderEditUser
         )
         protected.post(
-            "/admin/users",
+            endpoints.adminPanelUserBasePath,
             U.parameter,
-            "edit",
-            use: config.controllers.adminPanelUserController.editUser
+            endpoints.editSlug,
+            use: controllers.adminPanelUserController.editUser
         )
         protected.post(
-            "/admin/users",
+            endpoints.adminPanelUserBasePath,
             U.parameter,
-            "delete",
-            use: config.controllers.adminPanelUserController.delete
+            endpoints.deleteSlug,
+            use: controllers.adminPanelUserController.delete
         )
         protected.get(
-            "/admin/users/me/edit",
-            use: config.controllers.adminPanelUserController.renderEditMe
+            endpoints.adminPanelUserBasePath,
+            endpoints.meSlug,
+            endpoints.editSlug,
+            use: controllers.adminPanelUserController.renderEditMe
         )
         protected.post(
-            "/admin/users/me/edit",
-            use: config.controllers.adminPanelUserController.editMe
+            endpoints.adminPanelUserBasePath,
+            endpoints.meSlug,
+            endpoints.editSlug,
+            use: controllers.adminPanelUserController.editMe
         )
 
         // MARK: Reset routes
